backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pymupdf==1.24.9
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
tenacity==9.0.0
tiktoken==0.7.0
google-generativeai==0.8.2